Code cleanup and added finally for the Try block

+++ src/Hassium/Functions/MathFunctions.cs
@@ -18,12 +18,12 @@ namespace Hassium
         }
         public static object Pow(object[] args)
         {
-            return (double)(Math.Pow(Convert.ToDouble(args[0]), Convert.ToDouble(args[1])));
+            return Math.Pow(Convert.ToDouble(args[0]), Convert.ToDouble(args[1]));
         }
 
         public static object Sqrt(object[] args)
         {
-            return (double)(Math.Sqrt(Convert.ToDouble(args[0])));
+            return Math.Sqrt(Convert.ToDouble(args[0]));
         }
 
         public static object Hash(object[] args)
